CuboidGrids

This file contains all the dependencies, the data structs used in the data process. Has been added some functions (bottom page) from the voxellization example file so the voxel/render files will eventually contains less strings, resulting in a much better looking code.

##Main interface

CuboidGrids.Points — Type

Points = Array{Number,2}

Alias declation of LAR-specific data structure. Dense Array{Number,2,1} $M x N$ to store the position of vertices (0-cells) of a cellular complex. The number of rows $M$ is the dimension of the embedding space. The number of columns $N$ is the number of vertices.

CuboidGrids.Cells — Type

Cells = Array{Array{Int,1},1}

Alias declation of LAR-specific data structure. Dense Array to store the indices of vertices of P-cells of a cellular complex. The linear space of P-chains is generated by Cells as a basis. Simplicial P-chains have $P+1$ vertex indices for cell element in Cells array. Cuboidal P-chains have $2^P$ vertex indices for cell element in Cells array. Other types of chain spaces may have different numbers of vertex indices for cell element in Cells array.

CuboidGrids.Chain — Type

Chain = SparseArrays.SparseVector{Int8,Int}

Alias declation of LAR-specific data structure. Binary SparseVector to store the coordinates of a chain of N-cells. It is nnz=1 with value=1 for the coordinates of an elementary N-chain, constituted by a single N-chain.

CuboidGrids.ChainOp — Type

ChainOp = SparseArrays.SparseMatrixCSC{Int8,Int}

Alias declation of LAR-specific data structure. SparseMatrix in Compressed Sparse Column format, contains the coordinate representation of an operator between linear spaces of P-chains. Operators $P-Boundary : P-Chain -> (P-1)-Chain$ and $P-Coboundary : P-Chain -> (P+1)-Chain$ are typically stored as ChainOp with elements in ${-1,0,1}$ or in ${0,1}$, for signed and unsigned operators, respectively.

CuboidGrids.ChainComplex — Type

ChainComplex = Array{ChainOp,1}

Alias declation of LAR-specific data structure. It is a 1-dimensional Array of ChainOp that provides storage for either the chain of boundaries (from D to 0) or the transposed chain of coboundaries (from 0 to D), with D the dimension of the embedding space, which may be either $R^2$ or $R^3$.

CuboidGrids.LARmodel — Type

LARmodel = Tuple{Points,Array{Cells,1}}

Alias declation of LAR-specific data structure. LARmodel is a pair (Geometry, Topology), where Geometry is stored as Points, and Topology is stored as Array of Cells. The number of Cells values may vary from 1 to N+1.

CuboidGrids.LAR — Type

LAR = Union{ Tuple{Points, Cells},Tuple{Points, Cells, Cells} }

Alias declation of LAR-specific data structure. LAR is a pair (Geometry, Topology), where Geometry is stored as Points, and Topology is stored as Cells.

CuboidGrids.approxVal — Function

approxVal(PRECISION)(value)

Transform the float value to get a PRECISION number of significant digits.

CuboidGrids.CV2FV — Function

CV2FV( Array{Int64} ) Array{Int64} Return the vertices cells array tranformed to a vertices faces array

CuboidGrids.K — Function

K( Array{Int64} ) Return the sparse matrix built with the passed array

Missing docstring.

Missing docstring for CuboidGrids.VEF. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CuboidGrids.Mats. Check Documenter's build log for details.