largrid

This file is the heart of the project. Itit contains the function used for the voxellization process. The parallelization has been implemented intially with the pmap function. Another attempt was to diassamble the the functions and parallelize them with the @thread macro. This attempt was almost istantly discarded because it would mean make the same function more complicated, adding computational time. The whole process start from the call of the cuboidGrid function, which only need the shape of the data that is wanted to be voxellized. From the shape will be extracted the vertices array and the cells array.

Function list

CuboidGrids.larVertProd — Function

larVertProd(vertLists::Array{Points,1})::Points

Generate the integer coordinates of vertices (0-cells) of a multidimensional grid. Grid n-vertices are produced by the larVertProd function, via Cartesian product of vertices of $n$ 0-dimensional arguments (vertex arrays in vertLists), orderly corresponding to $x_1, x_2, ..., x_n$ coordinates in the output points $(x_1, x_2,...,x_n)$ in $R^n$.

# Example

julia> larVertProd([ larGrid(3)(0), larGrid(4)(0) ])
# output
2×20 Array{Int64,2}:
 0  0  0  0  0  1  1  1  1  1  2  2  2  2  2  3  3  3  3  3
 0  1  2  3  4  0  1  2  3  4  0  1  2  3  4  0  1  2  3  4

CuboidGrids.larImageVerts — Function

larImageVerts(shape::Array{Int,1})::Array{Int64,2}

Linearize the grid of integer vertices, given the shape of a cuboidal grid (typically an image).

Examples

julia> larImageVerts([1024,1024])
# output
2×1050625 Array{Int64,2}:
 0  0  0  0  0  0  0  0  0  0   0   0 … 1024  1024  1024  1024  1024  1024  1024  1024
 0  1  2  3  4  5  6  7  8  9  10  11 … 1017  1018  1019  1020  1021  1022  1023  1024

julia> larImageVerts([1,1,1])
# output
3×8 Array{Int64,2}:
 0  0  0  0  1  1  1  1
 0  0  1  1  0  0  1  1
 0  1  0  1  0  1  0  1

CuboidGrids.filterByOrder — Function

filterByOrder( n::Int )Array{Array{Array{Int8,1},1},1}

Filter the array of codes (Boolean String) of $n$ bits depending on their integer value (order).

Example

julia> filterByOrder(3)
# output
4-element Array{Array{Array{Int8,1},1},1}:
 Array{Int8,1}[Int8[0, 0, 0]]
 Array{Int8,1}[Int8[0, 0, 1], Int8[0, 1, 0], Int8[1, 0, 0]]
 Array{Int8,1}[Int8[0, 1, 1], Int8[1, 0, 1], Int8[1, 1, 0]]
 Array{Int8,1}[Int8[1, 1, 1]]

CuboidGrids.larCellProd — Function

larCellProd(cellLists::Array{Cells,1})::Cells

Generation of grid cells by Cartesian product of 0/1-complexes. The output complex is generated by the product of any number of either 0- or 1-dimensional cell complexes. The product of $d$ 1-complexes generates solid $d$-cells, while the product of $n$ 0-complexes and $d-n$ 1-complexes ($n < d$) generates non-solid $(d-n)$-cells, properly embedded in $d$-space, i.e. with vertices having $d$ coordinates.

Examples

To understand the generation of cuboidal grids from products of 0- or 1-dimensional complexes, below we show a simple example of 2D grids embedded in $R^3$. In particular, v1 = [0. 1. 2. 3.] and v0 = [0. 1. 2.] are two 2-arrays of 1D vertices, c1 = [[0,1],[1,2],[2,3]] and c0 = [[0],[1],[2]] are the LAR representation of one $1$-complex and one $0$-complex, respectively. The solid 2-complex named grid2D is generated in 2D as follows:

julia> v1 = [0. 1. 2. 3.]
1×4 Array{Float64,2}:
 0.0  1.0  2.0  3.0

julia> c1 = [[0,1],[1,2],[2,3]]
3-element Array{Array{Int64,1},1}:
 [0, 1]
 [1, 2]
 [2, 3]

julia> grid2D = larVertProd([v1,v1]),larCellProd([c1,c1])
([0.0 0.0 … 3.0 3.0; 0.0 1.0 … 2.0 3.0], Array{Int64,1}[[1, 2, 5, 6], [2, 3, 6, 7], [3, 4, 7, 8], [5, 6, 9, 10], [6, 7, 10, 11], [7, 8, 11, 12], [9, 10, 13, 14], [10, 11, 14, 15], [11, 12, 15, 16]])

whereas a non-solid $2$-complex in $3D$ is generated as:

julia> v1, c1 = [0. 1. 2. 3.],[[0,1],[1,2],[2,3]]
([0.0 1.0 2.0 3.0], Array{Int64,1}[[0, 1], [1, 2], [2, 3]])

julia> v0, c0 = [0. 1. 2.], [[0],[1],[2]]
([0.0 1.0 2.0], Array{Int64,1}[[0], [1], [2]])

julia> vertGrid = larVertProd([v1, v1, v0])
3×48 Array{Float64,2}:
 0.0  0.0  0.0  0.0  0.0  0.0  …  3.0  3.0  3.0  3.0  3.0  3.0  3.0  3.0  3.0
 0.0  0.0  0.0  1.0  1.0  1.0  …  1.0  1.0  1.0  2.0  2.0  2.0  3.0  3.0  3.0
 0.0  1.0  2.0  0.0  1.0  2.0  …  0.0  1.0  2.0  0.0  1.0  2.0  0.0  1.0  2.0

julia> cellGrid = larCellProd([c1, c1, c0])
27-element Array{Array{Int64,1},1}:
 [1, 4, 13, 16]
 [2, 5, 14, 17]
 ...  ... ...
 [32, 35, 44, 47]
 [33, 36, 45, 48]

julia> grid3D = vertGrid,cellGrid
([0.0 0.0 … 3.0 3.0; 0.0 0.0 … 3.0 3.0; 0.0 1.0 … 1.0 2.0], Array{Int64,1}[[1, 4, 13, 16], [2, 5, 14, 17], … [32, 35, 44, 47], [33, 36, 45, 48]])

julia> using Plasm

julia> Plasm.view(grid3D)

CuboidGrids.larGridSkeleton — Function

larGridSkeleton( shape::Array{Int,1} )( d::Int )::Cells

Produce the d-dimensional skeleton (set of d-cells) of a cuboidal grid of given shape.

Example

A shape=[1,1,1] parameter refers to a grid with a single step on the three axes, i.e. to a single 3D unit cube. Below all skeletons of such simplest grid are generated.

julia> Lar.larGridSkeleton([1,1,1])(0)
# output
8-element Array{Array{Int64,1},1}:
[[1], [2], [3], [4], [5], [6], [7], [8]]

julia> Lar.larGridSkeleton([1,1,1])(1)
# output
12-element Array{Array{Int64,1},1}:
[[1,2],[3,4],[5,6],[7,8],[1,3],[2,4],[5,7],[6,8],[1,5],[2,6],[3,7],[4,8]]

julia> Lar.larGridSkeleton([1,1,1])(2)
# output
6-element Array{Array{Int64,1},1}:
[[1,2,3,4], [5,6,7,8], [1,2,5,6], [3,4,7,8], [1,3,5,7], [2,4,6,8]]

julia> Lar.larGridSkeleton([1,1,1])(3)
# output
1-element Array{Array{Int64,1},1}:
 [1, 2, 3, 4, 5, 6, 7, 8]

CuboidGrids.cuboidGrid — Function

cuboidGrid( shape, filled=false )::Union( Cells, Array{Cells,1} )

Multi-dimensional generator function. Generate either a solid $d$-grid of unit $d$-cuboids in $d$-dimensional space, or the array of $p$-skeletons ($0 <=p<= d$), depending on the Boolean variable filled. $0$-cuboids are points, $1$-cuboids are segments, , $2$-cuboids are squares, $3$-cuboids are cubes, etc. The shape=[a,b,c] value determines the number $a x b x c$ of $d$-cells. Notice that d = length(shape)